requirements.txt
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google
streamlit
python-dotenv